Continue to try to make it work
There seems to be no way of creating a body in the HTTP request with the
the library I picked. The value that is sent to AdafruitIO must be in
the body (not in the header like the API key). Wether I add the
functionnality in the library or I find another one